Fix trunctation warnings in VGUI Grid

+++ vgui_utils/vgui_grid.cpp
@@ -321,10 +321,10 @@ void CGrid::SetRowUnderline(int row, bool enabled, int offset, int r, int g, int
 	if (enabled)
 	{
 		cell->m_iUnderlineOffset = offset;
-		cell->m_UnderlineColor[0] = r;
-		cell->m_UnderlineColor[1] = g;
-		cell->m_UnderlineColor[2] = b;
-		cell->m_UnderlineColor[3] = a;
+		cell->m_UnderlineColor[0] = static_cast<short>( r );
+		cell->m_UnderlineColor[1] = static_cast<short>( g );
+		cell->m_UnderlineColor[2] = static_cast<short>( b );
+		cell->m_UnderlineColor[3] = static_cast<short>( a );
 	}
 }
